requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.13
langchain-groq==0.2.0
langsmith==0.1.137
opentelemetry-api==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.10
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
zipp==3.20.2
